Add inner test module for simplifying setup.
